+++ src/main/java/com/jeremyseq/inhabitants/blocks/ModBlocks.java
@@ -2,10 +2,12 @@ package com.jeremyseq.inhabitants.blocks;
 
 import com.jeremyseq.inhabitants.Inhabitants;
 import com.jeremyseq.inhabitants.items.ModItems;
+import com.jeremyseq.inhabitants.blocks.impaler_head.ImpalerHeadBlock;
+
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.*;
+
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
@@ -20,6 +22,16 @@ public class ModBlocks
     public static final RegistryObject<Block> INVISIBLE_CAULDRON_BLOCK = registerBlock(
             "invisible_cauldron_block",
             () -> new InvisibleCauldronBlock(Block.Properties.of().noLootTable().strength(-1.0F, 3600000.0F).noOcclusion())
+    );
+
+    public static final RegistryObject<Block> IMPALER_HEAD = registerBlock(
+            "impaler_head",
+            () -> new ImpalerHeadBlock(Block.Properties.of().strength(1.5F, 1.5F).noOcclusion())
+    );
+
+    public static final RegistryObject<Block> DRIPSTONE_IMPALER_HEAD = registerBlock(
+            "dripstone_impaler_head",
+            () -> new ImpalerHeadBlock(Block.Properties.of().strength(1.5F, 1.5F).noOcclusion())
     );
 
     private static <T extends Block> RegistryObject<T> registerBlock(String name, Supplier<T> block) {