+++ src/main/java/com/jeremyseq/inhabitants/particles/ModParticles.java
@@ -47,6 +47,10 @@ public class ModParticles {
             PARTICLES.register("abracadabra",
                     () -> new SimpleParticleType(false));
 
+    public static final RegistryObject<SimpleParticleType> IMPALER_HEAD_ATTACK =
+            PARTICLES.register("impaler_head_attack",
+                    () -> new SimpleParticleType(false));
+
     @SubscribeEvent
     public static void registerParticles(RegisterParticleProvidersEvent evt) {
         evt.registerSpriteSet(ModParticles.IMPALER_SCREAM.get(), ScreamParticle.Factory::new);
@@ -57,5 +61,6 @@ public class ModParticles {
         evt.registerSpriteSet(ModParticles.SONIC_WAVE.get(), SonicWaveParticle.Factory::new);
         evt.registerSpriteSet(ModParticles.SHOCKWAVE.get(), ShockwaveParticle.Factory::new);
         evt.registerSpriteSet(ModParticles.ABRACADABRA.get(), AbracadabraParticle.Factory::new);
+        evt.registerSpriteSet(ModParticles.IMPALER_HEAD_ATTACK.get(), ImpalerHeadAttackParticle.Factory::new);
     }
 }